perf: indexer score GEMM accumulate fp32→bf16 (vllm#53459)

+++ vllm/models/deepseek_v4/attention.py
@@ -599,7 +599,7 @@ class DeepseekV4Attention(nn.Module, AttentionLayerBase, ABC):
                 return torch.mm(
                     hidden_states,
                     indexer.compressor.fused_wkv_wgate.weight.T,
-                    out_dtype=torch.float32,
+                    out_dtype=torch.bfloat16,
                 )
 
             aux_fns[1] = indexer_weights_proj
